refactor: rename `services` to `router`
