+++ views_api.py
@@ -14,7 +14,7 @@ from .crud import add_relay, delete_relay, get_relays
 from .helpers import normalize_public_key
 from .models import Relay, RelayList, TestMessage, TestMessageResponse
 from .nostr.key import EncryptedDirectMessage, PrivateKey
-from .services import NostrRouter, nostr
+from .router import NostrRouter, nostr
 
 # we keep this in
 all_routers: list[NostrRouter] = []
